--- dynamixel_ros_control/test/config/controller_spawner.yaml
@@ -1,49 +1,0 @@
-/**:
-  ros__parameters:
-    # How long to wait between retry attempts (HW + controllers).
-    retry_delay: 5.0
-
-    # ---------- hardware interfaces (activate = always) ----------
-    hardware_interfaces:
-      - athena_flipper_interface
-      - athena_arm_interface
-
-    # ---------- controllers --------------------------------------
-    controllers:
-      #- joint_state_broadcaster
-      #- flipper_position_controller
-      #- flipper_velocity_controller
-      #- flipper_trajectory_controller
-      - arm_position_controller
-      #- arm_velocity_controller
-      #- arm_trajectory_controller
-      #- gripper_position_controller
-      #- gripper_trajectory_controller
-
-    # ---------- Per‑controller options ---------------------------
-
-    # Always active to provide oint data
-    joint_state_broadcaster:
-      activate: true
-
-    # FLIPPERS: Starting with Velocity Control active
-    flipper_velocity_controller:
-      activate: true
-    flipper_position_controller:
-      activate: false
-    flipper_trajectory_controller:
-      activate: false
-
-    # ARM: Starting with Position Control active
-    arm_position_controller:
-      activate: true
-    arm_velocity_controller:
-      activate: false
-    arm_trajectory_controller:
-      activate: false
-
-    # GRIPPER: Starting with Position Control active
-    gripper_position_controller:
-      activate: true
-    gripper_trajectory_controller:
-      activate: false
